test/5_rest_api/requirements.txt
asgiref==3.8.1
# Django==5.0.4
django
django-filter==24.2
djangorestframework==3.15.1
sqlparse==0.4.4
typing_extensions==4.11.0
